requirements.txt
Jinja2==2.7.3
Mako==1.0.0
MarkupSafe==0.23
Pygments==2.0.1
SQLAlchemy==0.9.8
alembic==0.7.1
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
certifi==14.05.14
fancycompleter==0.4
-e git+git@github.com:quantopian/ipython.git@8fd1a6f42bd2785c902ab0411b50afa4051699b1#egg=ipython-master
jsonschema==2.4.0
mistune==0.5.0
nose==1.3.4
pdbpp==0.7.2
psycopg2==2.5.4
pyrepl==0.8.4
pyzmq==14.4.1
tornado==4.0.2
wmctrl==0.1
wsgiref==0.1.2
flake8==2.2.5
requests==2.5.0
six==1.8.0
click==3.3
python-dateutil
